removing filter
