+++ assert/main.go
@@ -1,0 +1,17 @@
+package assert
+
+import (
+	"os"
+	"testing"
+)
+
+// Main wraps a testing.M.
+func Main(m *testing.M) {
+	Started()
+	var statusCode int
+	func() {
+		defer ReportRate()
+		statusCode = m.Run()
+	}()
+	os.Exit(statusCode)
+}
